Command interpretation update, general cleanup
Command interpretation replaced with a loop for the eventual full move to command interpretation loop (?). Small edit made to the way message content is interpreted, abstracting it into a secondary, editable string in order to facilitate manipulation of said string in commands. (this change was not tested, test it before you push it thanks)